Remove sync_to_tk representability checks (#16407)
These checks verified that the current YAML config could be exactly reproduced
using TK settings. That made sense back when TK might actually be used for
scheduling, but it doesn't make sense now: we can fully control the behavior of
the constellation with the channel configs as long as the _required_ settings
are all turned on. It doesn't matter if YAML-disabled channels are enabled in
TK, but all of the YAML-enabled ones must also be on in TK for the system to
work. The current checks are overly strict, so I am removing them for now while
I check to see if we can take a different approach.