Add PITGS STA constraints to staging satellites (#17936)
* Add fragment definitions for PITGS STA constraints

These fragments define constraints required by our Special Temporary
Authority to operate the Pittsburgh ground station. Unfortunately, one
of them is a bit broad and prevents us from using the 2022.5 MHz bands
from any of our satellites while the ISS is within 150km. The other
prevents UHF comms between any satellite and PITGS when the ISS is
within view of the ground station.

See: https://docs.google.com/document/d/1oGeO3ofSt-Hd-UTpnYScEQ4EcrycYDSq/edit

* Move FM123 constraint comment inside the YAML block

This should prevent it from being lost when new constraints are merged
in. The comments are preserved as part of the Ruamel YAML data
structures, but in their previous position they were part of the list
structure, not the dictionary that a human would expect. When new
elements are added to the sequence and it is converted to a Python list
or set the comments can be discarded inadvertently.

* Apply PITGS STA S-band constraints to staging sats

* Apply PITGS STA UHF constraints to staging sats