feat: add inspections to channel_tool audit (#21782)
* refactor: reorganize channel_tool module layout

The tool is getting quite large, with lots of functionality all cobbled
together into one mega-module. This makes it hard to work with and
refactor. As a first step towards enabling easier edits, this commit
moves the tool and supporting libraries into a sub-module and updates
the associated tests.

* refactor: extract audit, tk, and asset_config modules

These are coupled modules that needed to be pulled out together so that
I can have the `audit` module separated for extension.

* refactor: extract AuditReport class

This is a step towards being able to have different annotations in audit
output, beyond just the "Subject to satellite constraints" note.

* refactor: extract SatelliteConstraintInspection

This extracts the first of several inspections that can be added to an
AuditReport, adding the wiring so that we can run them.

* feat: add headers to inspection output

* feat: add MergedParametersInspection showing static params

* feat: add contact type inspection

Channels can provide a contact type override. This should either match
on both the satellite and ground station or be set on at most one of
them. This inspection shows which one is being used.

* style: add type annotations to help MyPy

* fix: use new module invocation in ci script

* docs: update readme

* refactor: replace if/else chain with match