Migrate the CI pipeline from concourse CI to Github Action (#424)
Migrate the CI from Concourse CI to Github Actions:

Remove the ci/ directory, the ci pipeline in concourse will be removed after this PR is merged
Add the github workflow files:
check.yml define the setup job that will prepare a .venv/ directory with poetry and the dependencies. This directory will be cached between runs. The other jobs (formatter, behave, pytest, mypy, channel validate) will be run in parallel from the cached .venv/
problem_matchers: Strings used to highlight the errors in mypy and black
While debugging an unrelated issue, I found that the tests in features/basic_edits.feature weren't checking that the commands were run successfully and would silently ignore failures. I made a small change to the test definition to fix that.

Updated isort config to make it generate black-compatible files.

Features:
- Error detection in mypy and black's logs
- Code coverage and Tests summary for Pytest

This PR is a follow-up of #427 which fixed the formatting of the python files.