Classification annotations, template mechanism reform, and other small updates (#444)
**Classification annotations**

This PR adds `classification_annotations` as a new mandatory top-level
element to each groundstation's channel config. It is a set of flat
key-value pairs which will be added to the contact's annotations in TK
(after the appropriate corresponding optimizer repo feature is
deployed). These classification annotations are designed as a public API
to help any system consuming the contact schedule to filter on different
types of contact.

A small feature was added to the `channel_tool` to allow this to be
populated from a file, and fragments were added for all current
channels. The actual config changes were done with this piece of `bash`

```
CHANNELS=`grep -oh ^\[0-9A-Z_\]\*:$ {production,staging}/{sat,gs}/*.yaml | sed s/://g | sort -u`

for CHANNEL in $CHANNELS; do
    poetry run python -m channel_tool --debug edit -y --classification_annotations_file fragments/classification_annotations/$CHANNEL.yaml staging all_gs $CHANNEL
    poetry run python -m channel_tool --debug edit -y --classification_annotations_file fragments/classification_annotations/$CHANNEL.yaml production all_gs $CHANNEL
    done
```

Logs here for an example of running this feature:
https://paste.ect.spire.com/evojobaren.bash

**Schema update**

The PR also makes an unrelated schema change, allowing the
recently-added `additional_provider_config` to occur on the
groundstation side as well as on the satellite side. There will need to
be an optimizer PR to make use of this.

The `chainOverride` element was removed from the schema and configs, it
is no longer used / needed.

**Templates mechanism**

We also updated the templates mechanism. It was not being used and has
not been kept sufficiently up-to-date. Also the idea that you would have
the same structure (and therefore template) for a channel on both the GS
and satellite side turned out not to be correct. Specifically the PR
makes the following changes:
- removed the `templates.yaml` file, replacing it with
`gs_templates.yaml` and `sat_templates.yaml`
- updated the contents of these files to contain the currently in-use
channels
- in the schema, clearly separate the GS from the satellite schema.
- more / updated tests
- validation checks that configs only use channels in the relevant
template
- validation checks that the gs template file has the same set of
channels as the sat template file
- validation checks that all channels in template files have legal:
false and enabled: false

Co-authored-by: John Toop-Rose <john.tooprose@gcc.spire.com>