Disabling DVB due to OP-7648 (#571)
OP-7648 is blocking C&C progress - [C&C] OORT Transport/Dexter/DVB
Download issues -

This PR disables DVB, as it seems like the bug found is mainly present
in the DVB RXOs. This is just a workaround to have more RXO success, so
that we can resume C&C and also get more logs to troubleshoot OP7648,
but eventually we will have to revert this PR to ensure we are achieving
nominal download performance.

The reason sband legacy works (a bit better) is that the dvb_s2x has
that 10s warmup which is exacerbating this race condition.