[staging] Override bandaid_override.oort_dest_host to override a bad value in bandaid config (#604)
The current version of the bandaid config `dexter_v3.json` has an
invalid value for `oort_dest_host` (marsupial dexter). This becomes
visible on sband up contacts.
This PR will override the value during the contacts. This PR is only on
staging for now to make sure that it is working as intended