missing sun constraint added to the new X/S bidir channels on 3 x GHGSat FMs (#627)
Looks like we forgot to put the sun constraint on the new X/S bidir
channels for GHGSat, so they are getting scheduled in sunlit periods,
contrary to the requirements. @nick-molo and Neil Aitken confirmed to me
that there has not been any special agreement around this with GHGSat,
so we should just apply the sun constraint to these new channels.