evolution of the "validation rules" feature (#639)
Cleanup / restructure of how validation rules work.

- 'validate' command in channel tool has 2 new CLI options, that allow
the rules used to be restricted by module and function name
- utils module is removed, with stuff moving up to validation_rules.
This is because we don't want utils to look like a rules module
- rules are written using the decorator @validation_rule, this allows
common stuff, e.g. relating to the scope of the rules, to be factored
out. The rule then just returns a boolean, true for pass, false for
fail.
- various tweaks around how rules are run and logged
- updates to the concrete classification annotation rules, to correct
these.