apply staging electrical cost to production for 194 and 196 (#646)
With the staging for several days with the configs in place, it appears
we shouldn't expect to lose any contacts on this first iteration, and in
fact schedule many more contacts of all types in staging. Thus, it can't
be because of charge that we would reduce contact count in production as
there is enough charge for contacts of all types in staging relative to
production. Thus we can verify things work in production and proceed
with the experiment by subsequently dropping the constraints iteratively
in production TK