SBand up to DE-licensed satellites from awags is not allowed (#651)
Unfortunately I think we have made an error during the rollout of SBand
up. AWAGS is not actually licensed to use SBand up with DE-licensed FMs
(178,179,180,182,184,185,186,187,188,189,190). The error was in the
PLS29 / tracking channel. On one of the other X/S bidir channels
@kyle-henry correctly excluded NZ.

This grafana link shows the specific times when we used AWAGS ->
DE-licensed SBand up.

https://grafana-ect.cloud.spire.com/d/4c6IuLh4z/scheduled_contacts_production?orgId=1&var-groundstation=awags&var-satellite=FM178&var-satellite=FM179&var-satellite=FM180&var-satellite=FM182&var-satellite=FM184&var-satellite=FM185&var-satellite=FM186&var-satellite=FM187&var-satellite=FM188&var-satellite=FM189&var-satellite=FM190&var-channel_id=All&var-provider=All&var-directionality=All&var-adcs_pointing=All&var-space_ground_uhf=All&var-space_ground_sband=All&var-space_ground_xband=All&var-ground_space_sband=true&from=now-180d&to=now