UHF Only Operations (#665)
Editing channel configuration as well as TK to try and have the station
capable of producing some beneficial UHF contact time while we sort
through the issue of a broken rotor. Rotor is parked in birdbath so we
have increased the elevation angle and shortened the overhead time in an
effort to get some contacts scheduled utilizing the UHF large antenna
beamwidth.