Updates to enforce PLS consistency within a channel (#680)
PLS value is fundamental part of a channel definition. So we should not
be overriding PLS to different values within a channel.

- Introduce 4 new validation rules to enforce that any PLS overriding in
bandaid overrides, either on the sat or GS side, is consistent with the
declared PLS value in the classification annotations for the channel.
- Fix cases where the PLS rule was violated:
CONTACT_RXO_DVBS2X_SBAND_FREQ_2200_MHZ on ubngs, awags and pergs, where
PLS was overridden to 87 from the default 39. We introduce a new channel
which is the same but has PLS 87.