Generate optimal channels for each asset pair (#704)
This PR introduces a script that can, using data from Project Ripley,
generate/maintain channels for each asset pair such that for the given
PLS/elevation, we know the pair can support the required SNR 95% of the
time.

Note that if a satellite currently has it's DVB S-Band channel disabled,
it doesn't get the new channels, even in disabled state.

This PR will be followed by a staging PR and a production PR with the
generated channels.

The experiment is tracked in
[RPI-211](https://jira.spire.com/browse/PRI-211).

### Current channel configuration method

We create a few channels with chosen PLS/BW values and assign the same
channel to all assets that can support the configuration in theory
(hardware compatibility, legality, etc.). The asset a channel belongs to
plays little to no role in it's link characteristics.

e.g. `CONTACT_RXO_DVBS2X` is a 39PLS, 1MHZ channel that is configured to
work above 25 degrees on all assets it's added to.

### Does it work?

In practice we see a great variation across not only individual assets
but asset pairs in terms of SNR performance in the data. This is
triggered by hardware design, antenna patterns, etc. The data suggests
that most assets are configured to do contacts they aren't capable of in
terms of ZBC, bitrate, and SNR performances.

##### p95 SNR by asset pair in 30-35 degrees elevation range
[source](https://dbc-24e0a945-16d8.cloud.databricks.us/dashboardsv3/01efdd9bbb391244a044494011a63f17/published/pages/01efeebcb6881f5096c6b6499ed9ae6a?o=875599787854343&f_01efe21218a010fbb9a3bbd6925c23a6=%257B%2522columns%2522%253A%255B%2522x%2522%255D%252C%2522rows%2522%253A%255B%255B%252287%2522%255D%255D%257D&f_01efdd9bbb421fddbddc76531751f64f=%257B%2522columns%2522%253A%255B%2522x%2522%255D%252C%2522rows%2522%253A%255B%255B%2522awags%2522%255D%255D%257D&f_01efe48df4a91f8cba4ca1a7063ca439=awags&f_01efe4a750351d89aaa676c03194da0d=CONTACT_RXO_DVBS2X_SBAND_FREQ_2200_MHZ&f_01efdd9bbb3d1e6a957250755b70dfc5=hlegs&f_01efdd9bbb3d1e938dd11664abb277c0=FM131&f_01efed1f6d8d1f4a9f4cc4f82623a1d8=_all_&f_01efdd9bbb3d1d1b8caa370e3fbf0616=e6008f8d-a63c-4297-9a09-eeb2ef781415&f_01efeec122381be1b602214dd0f6c675=25)


![image](https://github.ect.spire.com/storage/user/445/files/34aaa3c9-170c-46f2-b169-2cff49359458)

### What does the script do?

1. Pull the relevant
[data](https://dbc-24e0a945-16d8.cloud.databricks.us/explore/data/ripley_dev/default/optimal_channel_properties_gs_fm_3_db_buffer?o=875599787854343&activeTab=sample)
from Databricks (requires access to Ripley catalogs and a Databricks
token). The data lists all the achievable PLS values, the elevation at
which they can be achieved, and some details about the link (band,
bandwidth, bitrate, etc.) for all asset pairs.
2. For each configuration, it will find an already existing channel in
the assets. To do so, it uses the classification annotations.
Specifically, it will look for [these
fields](https://github.ect.spire.com/space/channel-config/blob/75aad4d607a201ce2a1cc057f58d022b1999cdf2/generate_optimal_channels.py#L237).
3. After finding an existing channel, it will duplicate that channel
within the asset and update certain values. This is to ensure we
preserve asset-specific configurations (licenses, constraints, etc.).
The list of updated fields are as follows:
* Contact overhead time is set to a flat 10s in case the existing
channel have it wrong.
* `jira_ticket` classification annotation.
* `space_ground_[s|x]band_dvbs2x_pls` classification annotation.
* `space_ground_[s|x]band_bandwidth_mhz` classification annotaion.
* `min_elevation_deg` and `downlink_rate_kbps` in the DVB link profile.
* Relevant fields in the DVB forward channel in `window_parameters`,
using output from the `pls` command.
* `dynamic_window_parameters.transmit_times.elevation_threshold_deg` is
updated to match the link profile.

### Misc changes

* Added `raw` switch to the `pls` subcommand that silences additional
commentary and outputs raw YAML bytes.

### Future work

* Move the script into channel tool as `duplicate` command.
* Improve update support (i.e. further maintaining generated channels).
* Add support for S-Band-up BIDIRs.