FM194 and FM196 fixes for UHF power cost (#753)
It was identified that these first two satellites that had charge
constraints applied have an inaccurate values being used for the power
used by the UHF radio. This PR applies the values that are on the
staging configurations to the production configurations, which is based
on data obtained from telemetry analysis.