predicate support as an alternative to channel names (#751)
This PR is adding another option to the `channels` positional argument.
Instead of specifying channel names, we'll allow specifying predicates
to filter channels dynamically based on classification annotations.

It is also porting pre-defined channel groups to use these predicates
instead. This means the groups will not go out-of-date and are easier to
create.