feat(gs): allow specifying different minimum elevations per satellite (#759)
This PR introduces a new optional schema for expressing link profiles.
The old
schema is still supported, however, they cannot be mixed in the same
channel.

# New Schema

The new schema introduces a `satellite_min_elevations` field to GS link
profiles. This field holds a mapping from a list of satellites to the
`min_elevation_deg` to be used for those satellites. The old
`min_elevation_deg` field becomes `default_min_elevation_deg` which now
holds
the value to be used if there was no satellite specific override is
found.

It also removes the need for
`dynamic_window_parameters.transmit_times.elevation_threshold_deg`.
Instead,
when this is required, `add_transmit_times` field in the relevant link
profile
must be set to true.

# Guarantees

We guarantee that:

- `min_elevation_deg` and `default_min_elevation_deg` fields are
mutually exclusive.
- `add_transmit_times` field requires `default_min_elevation_deg` field.
- `satellite_min_elevations` field requires `default_min_elevation_deg`
field.

These guarantees ensure that the new and old schemas cannot be mixed in
the
same channel. They should also make the transition easier.

This is part of the implementation of
https://confluence.spire.com/x/1YEhBw.