feat: duplicate command in channel tool (#758)
Duplicate a channel within an asset with a different PLS value.
Duplicate code is in a separate module. Non-DVB channels and S/X BIDIRs
cannot be duplicated for now, there's a function to check if `duplicate`
is supported for a given channel. PLS tool is slightly modified to
return what it's printing and allows supressing it's stdout. Useful for
calling it internally.

Note that channels can't be duplicated across assets. This is not a
current limitation but a design decision. Duplication functionality is
intended for preserving asset specific settings when creating a new
channel.

If more than one channel to duplicate is specified, only the first one
will be used. This is needed to allow using predicates.

---------

Co-authored-by: Josh Singer <josh.singer@gcc.spire.com>