Populate per-pair elevations in link profiles (#766)
This PR adds a script that can populate per-pair elevation fields in
link
profiles based on data it fetches from Databricks.

The script will be the ancestor of the eventual channel tool command to
do
this. Since it's temporary, correctness is the only priority.

The script currently only supports manipulating TXO channels.

I have fixed some issues with `duplicate` too:

- `duplicate_config` does not pick the first channel it finds when
multiple are
specified (e.g. when a predicate is specified). This is because it
cannot
know whether that channel will exist on the assets being operated on.
This is
a little inefficient as we'll end up with iterating over all found
channels -
attempting to duplicate all the ones we find on the asset. Duplication
is
  idempotent so this doesn't affect the results.
- `duplicate_config` does not generate a name for a channel now because
it does
  not know which channel will end up being duplicated.
- `apply_update` now accepts `new_channel` rather than `new_channel_id`.
It
  will generate the appropriate channel id.
- `duplicate_link_profile` populates either `min_elevation_deg` or
`default_min_elevation_deg` depending on which one exists in the
original
  link profile.