SMAGS channels with per-pair elevations (staging) (#790)
Creates S-Band TXO and BIDIR channels in SMAGS and all (relevant)
satellites in staging. The channels contain per-pair elevations. We have
4 supported values: 7, 15, 23, and 27. The previous channels with
unsupported PLS values are deleted from SMAGS (but not the satellites).

The population script is overhauled to address some edge cases I
encountered.

There's also a bugfix to duplicate command around window parameters when
there's no window parameter in the channel duplicated (i.e. when the
source was using the defaults).

See `run.log` for what commands were run. The files referenced by the
commands are also in the archive.


[Archive.zip](https://github.ect.spire.com/space/channel-config/files/1342/Archive.zip)