Fix forward channel orders for window parameters of SBAND BIDIRs (#809)
Optimizer assumes the order of forward channels on the satellite and the
groundstation are the same when merging them. This causes invalid window
parameter annotations when the order is different. This PR fixes the
order
for all assets and adds two validation rules that:
- Ensures the first item is always UHF and the second item is always
SBAND for SBAND BIDIRs.
- Ensures there are exactly 2 and 1 forward and reverse channels for
SBAND BIDIRs respectively.