FM229 production schedule DEU license (#806)
creating a schedule file for FM229 based on similar DEU licensed sat.
copy of FM216 as another DEU license. 
adopting sat constraints from FM226 as that has a matching dexter
module. (lemdig)

[MERGED] this PR requires
https://github.ect.spire.com/space/channel-config/pull/821
halftrack is added to each of the contact names when performing sband
since the patch antenna can be blocked by the lacuna antenna. sat will
do a nadirpoint and then a target track toward GS so that its signal is
not inturrupted.

convention to get configured satellite-side of configs, is to add
_HALFTRACK to contact types and then to add an adacs config. in this
case, we require
``` 
window_parameters:
    adcs_config:
        primary_instrument: LACUNA_NADIR
        secondary_instrument: SBAND_CONTACT
        mode: NADIRPOINTLATLON
```
note that mode is a config that is added on the GS side of the config,
and not technically necessary here but is expected by the validator.

---------

Co-authored-by: Zackary Bowles <zack.bowles@gcc.spire.com>