SNR autotune: AWAGS X-Band (staging) (#826)
Adds S/X BIDIR support to duplicate and creates the optimized channels
for
AWAGS.

> (gs_id = "awags" and band = 'X' and pls in (13, 25, 37, 57) and bw_mhz
= 10)

Some notes on S/X BIDIR duplication:
- Duplicate always writes a forward channel to window parameters even if
the
original channel did not have one.
- It uses the S-Band reverse channel as is from the original and does
not
generate a new one if the original did not have one. This is because
there's no
way to programatically generate an S-Band reverse channel window
parameter.

This is OK because S-Band reverse channels uses legacy encoding and
their
properties are not updated as part of duplicating a channel.

As a result S/X BIDIR channels on satellites have a forward channel but
not a
reverse channel. This is not an issue since the ground-side will have
the
reverse channel.

It looks like the S-Band reverse channels are the same across the board
but I
opted not to hardcode those properties into duplicate.

Also adds the `--bidir` flag to pls tool. Defaults to false. When given,
it'll
use `RX_SBAND_TX_XBAND` as its bandaid mode rather than
`TX_XBAND_DVB_IP` for
X-Band.