disabling xband channels on sg157kl due to increased x scheduling and network issue (#873)
sg157kl has been one of the newer batch of antennas opened up to us.

This new batch of antennas were all failing xband, believed due to a
networking issue on ksat side.

We disabled all the new batch except this one to keep some data points
coming in for investigation by ksat.

With increased weekend scheduling and ksat budgeting the schedule rate
has gone up here resulting in 70-80% zbc rate. This is a bit high to
accept for testing purposes and it may be best sats trying xband save
the power for another antenna as we try to perform well over the
weekend.